Revert "[datadog_user] Add service account support (#1490)"
This reverts commit 26bab4cc94f391277306cadf08d67da47f3860cb.
