--- datadog/tests/cassettes/TestAccDatadogUser_ServiceAccount.freeze
@@ -1,1 +1,0 @@
-2022-07-13T15:13:30.263289-07:00
