--- datadog/tests/cassettes/TestAccDatadogUser_ServiceAccount.yaml
@@ -1,84 +1,0 @@
----
-version: 1
-interactions:
-- request:
-    body: |
-      {"data":{"attributes":{"email":"tf-testaccdatadoguser_serviceaccount-local-1657750410@example.com","name":"Test User","service_account":true},"relationships":{"roles":{"data":[]}},"type":"users"}}
-    form: {}
-    headers:
-      Accept:
-      - application/json
-      Content-Type:
-      - application/json
-    url: https://api.datadoghq.com/api/v2/service_accounts
-    method: POST
-  response:
-    body: '{"data":{"type":"users","id":"eb284ef1-0e6b-43c9-94b6-3a1e406ac697","attributes":{"name":"Test User","handle":"eb284ef1-0e6b-43c9-94b6-3a1e406ac697","created_at":"2022-07-13T22:13:32.113345+00:00","modified_at":"2022-07-13T22:13:32.146709+00:00","email":"tf-testaccdatadoguser_serviceaccount-local-1657750410@example.com","icon":"https://secure.gravatar.com/avatar/f897c552e972b5ec6b9c7c64e8db4e67?s=48&d=retro","title":null,"verified":true,"service_account":true,"disabled":false,"allowed_login_methods":[],"status":"Active"},"relationships":{"roles":{"data":[]},"org":{"data":{"type":"orgs","id":"1bad3654-997a-11ec-844f-da7ad0900002"}}}}}'
-    headers:
-      Content-Type:
-      - application/json
-    status: 201 Created
-    code: 201
-    duration: ""
-- request:
-    body: ""
-    form: {}
-    headers:
-      Accept:
-      - application/json
-    url: https://api.datadoghq.com/api/v2/users/eb284ef1-0e6b-43c9-94b6-3a1e406ac697
-    method: GET
-  response:
-    body: '{"data":{"type":"users","id":"eb284ef1-0e6b-43c9-94b6-3a1e406ac697","attributes":{"name":"Test User","handle":"eb284ef1-0e6b-43c9-94b6-3a1e406ac697","created_at":"2022-07-13T22:13:32.113345+00:00","modified_at":"2022-07-13T22:13:32.146709+00:00","email":"tf-testaccdatadoguser_serviceaccount-local-1657750410@example.com","icon":"https://secure.gravatar.com/avatar/f897c552e972b5ec6b9c7c64e8db4e67?s=48&d=retro","title":null,"verified":true,"service_account":true,"disabled":false,"allowed_login_methods":[],"status":"Active"},"relationships":{"roles":{"data":[]},"org":{"data":{"type":"orgs","id":"1bad3654-997a-11ec-844f-da7ad0900002"}}}}}'
-    headers:
-      Content-Type:
-      - application/json
-    status: 200 OK
-    code: 200
-    duration: ""
-- request:
-    body: ""
-    form: {}
-    headers:
-      Accept:
-      - application/json
-    url: https://api.datadoghq.com/api/v2/users/eb284ef1-0e6b-43c9-94b6-3a1e406ac697
-    method: GET
-  response:
-    body: '{"data":{"type":"users","id":"eb284ef1-0e6b-43c9-94b6-3a1e406ac697","attributes":{"name":"Test User","handle":"eb284ef1-0e6b-43c9-94b6-3a1e406ac697","created_at":"2022-07-13T22:13:32.113345+00:00","modified_at":"2022-07-13T22:13:32.146709+00:00","email":"tf-testaccdatadoguser_serviceaccount-local-1657750410@example.com","icon":"https://secure.gravatar.com/avatar/f897c552e972b5ec6b9c7c64e8db4e67?s=48&d=retro","title":null,"verified":true,"service_account":true,"disabled":false,"allowed_login_methods":[],"status":"Active"},"relationships":{"roles":{"data":[]},"org":{"data":{"type":"orgs","id":"1bad3654-997a-11ec-844f-da7ad0900002"}}}}}'
-    headers:
-      Content-Type:
-      - application/json
-    status: 200 OK
-    code: 200
-    duration: ""
-- request:
-    body: ""
-    form: {}
-    headers:
-      Accept:
-      - '*/*'
-    url: https://api.datadoghq.com/api/v2/users/eb284ef1-0e6b-43c9-94b6-3a1e406ac697
-    method: DELETE
-  response:
-    body: ""
-    headers: {}
-    status: 204 No Content
-    code: 204
-    duration: ""
-- request:
-    body: ""
-    form: {}
-    headers:
-      Accept:
-      - application/json
-    url: https://api.datadoghq.com/api/v2/users/eb284ef1-0e6b-43c9-94b6-3a1e406ac697
-    method: GET
-  response:
-    body: '{"data":{"type":"users","id":"eb284ef1-0e6b-43c9-94b6-3a1e406ac697","attributes":{"name":"Test User","handle":"eb284ef1-0e6b-43c9-94b6-3a1e406ac697","created_at":"2022-07-13T22:13:32.113345+00:00","modified_at":"2022-07-13T22:13:33.515918+00:00","email":"tf-testaccdatadoguser_serviceaccount-local-1657750410@example.com","icon":"https://secure.gravatar.com/avatar/f897c552e972b5ec6b9c7c64e8db4e67?s=48&d=retro","title":null,"verified":true,"service_account":true,"disabled":true,"allowed_login_methods":[],"status":"Disabled"},"relationships":{"roles":{"data":[]},"org":{"data":{"type":"orgs","id":"1bad3654-997a-11ec-844f-da7ad0900002"}}}}}'
-    headers:
-      Content-Type:
-      - application/json
-    status: 200 OK
-    code: 200
-    duration: ""
