+++ datadog/tests/resource_datadog_application_key_test.go
@@ -76,74 +76,6 @@ func TestDatadogApplicationKey_import(t *testing.T) {
 	})
 }
 
-func TestDatadogApplicationKey_ServiceAccount(t *testing.T) {
-	if isRecording() || isReplaying() {
-		t.Skip("This test doesn't support recording or replaying")
-	}
-	t.Parallel()
-	resourceName := "datadog_application_key.foo"
-	ctx, accProviders := testAccProviders(context.Background(), t)
-	applicationKeyName := uniqueEntityName(ctx, t)
-	accProvider := testAccProvider(t, accProviders)
-
-	resource.Test(t, resource.TestCase{
-		PreCheck:          func() { testAccPreCheck(t) },
-		ProviderFactories: accProviders,
-		CheckDestroy:      testAccCheckDatadogApplicationKeyDestroy(accProvider),
-		Steps: []resource.TestStep{
-			{
-				Config: testAccCheckDatadogApplicationKeyServiceAccount(applicationKeyName, "foo"),
-				Check: resource.ComposeTestCheckFunc(
-					testAccCheckDatadogApplicationKeyExists(accProvider, resourceName),
-					resource.TestCheckResourceAttr(resourceName, "name", applicationKeyName),
-					testAccCheckDatadogApplicationKeyValueMatches(accProvider, resourceName),
-				),
-			},
-		},
-	})
-}
-
-func TestDatadogApplicationKey_ServiceAccountUpdateForceNew(t *testing.T) {
-	if isRecording() || isReplaying() {
-		t.Skip("This test doesn't support recording or replaying")
-	}
-	t.Parallel()
-	resourceName := "datadog_application_key.foo"
-	ctx, accProviders := testAccProviders(context.Background(), t)
-	applicationKeyName := uniqueEntityName(ctx, t)
-	accProvider := testAccProvider(t, accProviders)
-	applicationKeyNameUpdate := applicationKeyName + "-2"
-
-	resource.Test(t, resource.TestCase{
-		PreCheck:          func() { testAccPreCheck(t) },
-		ProviderFactories: accProviders,
-		CheckDestroy:      testAccCheckDatadogApplicationKeyDestroy(accProvider),
-		Steps: []resource.TestStep{
-			{
-				Config: testAccCheckDatadogApplicationKeyServiceAccount(applicationKeyName, "foo"),
-				Check: resource.ComposeTestCheckFunc(
-					testAccCheckDatadogApplicationKeyExists(accProvider, resourceName),
-					resource.TestCheckResourceAttr(resourceName, "name", applicationKeyName),
-					testAccCheckDatadogApplicationKeyValueMatches(accProvider, resourceName),
-				)},
-			{
-				Config: testAccCheckDatadogApplicationKeyServiceAccount(applicationKeyNameUpdate, "foo"),
-				Check: resource.ComposeTestCheckFunc(
-					testAccCheckDatadogApplicationKeyExists(accProvider, resourceName),
-					resource.TestCheckResourceAttr(resourceName, "name", applicationKeyNameUpdate),
-					testAccCheckDatadogApplicationKeyValueMatches(accProvider, resourceName),
-				)},
-			{
-				Config: testAccCheckDatadogApplicationKeyServiceAccount(applicationKeyNameUpdate, "bar"),
-				Check: resource.ComposeTestCheckFunc(
-					testAccCheckDatadogApplicationKeyExists(accProvider, resourceName),
-					resource.TestCheckResourceAttr(resourceName, "name", applicationKeyNameUpdate),
-					testAccCheckDatadogApplicationKeyValueMatches(accProvider, resourceName),
-				)},
-		},
-	})
-}
-
 func testAccCheckDatadogApplicationKeyConfigRequired(uniq string) string {
 	return fmt.Sprintf(`
 resource "datadog_application_key" "foo" {
@@ -151,33 +83,26 @@ resource "datadog_application_key" "foo" {
 }`, uniq)
 }
 
-func testAccCheckDatadogApplicationKeyServiceAccount(uniq, serviceAccount string) string {
-	return fmt.Sprintf(`
-data "datadog_role" "ro_role" {
-  filter = "Datadog Read Only Role"
+func testAccCheckDatadogApplicationKeyExists(accProvider func() (*schema.Provider, error), n string) resource.TestCheckFunc {
+	return func(s *terraform.State) error {
+		provider, _ := accProvider()
+		providerConf := provider.Meta().(*datadog.ProviderConfiguration)
+		apiInstances := providerConf.DatadogApiInstances
+		auth := providerConf.Auth
+
+		if err := datadogApplicationKeyExistsHelper(auth, s, apiInstances, n); err != nil {
+			return err
+		}
+		return nil
+	}
 }
 
-resource "datadog_user" "foo" {
-  email = "new@example.com"
-
-  roles = [data.datadog_role.ro_role.id]
-
-  service_account = true
-}
-
-resource "datadog_user" "bar" {
-  email = "new2@example.com"
-
-  roles = [data.datadog_role.ro_role.id]
-
-  service_account = true
-}
-
-resource "datadog_application_key" "foo" {
-  name = "%s"
-
-  service_account = datadog_user.%s.id
-}`, uniq, serviceAccount)
+func datadogApplicationKeyExistsHelper(ctx context.Context, s *terraform.State, apiInstances *utils.ApiInstances, name string) error {
+	id := s.RootModule().Resources[name].Primary.ID
+	if _, _, err := apiInstances.GetKeyManagementApiV2().GetCurrentUserApplicationKey(ctx, id); err != nil {
+		return fmt.Errorf("received an error retrieving application key %s", err)
+	}
+	return nil
 }
 
 func testAccCheckDatadogApplicationKeyValueMatches(accProvider func() (*schema.Provider, error), n string) resource.TestCheckFunc {
@@ -198,21 +123,12 @@ func datadogApplicationKeyValueMatches(ctx context.Context, s *terraform.State, 
 	primaryResource := s.RootModule().Resources[name].Primary
 	id := primaryResource.ID
 	expectedKey := primaryResource.Attributes["key"]
-	var last4OfKey string
-	if v, ok := s.RootModule().Resources[name].Primary.Attributes["service_account"]; ok {
-		resp, _, err := apiInstances.GetServiceAccountsApiV2().GetServiceAccountApplicationKey(ctx, v, id)
-		if err != nil {
-			return fmt.Errorf("received an error retrieving application key %s", err)
-		}
-		last4OfKey = resp.Data.Attributes.GetLast4()
-	} else {
-		resp, _, err := apiInstances.GetKeyManagementApiV2().GetCurrentUserApplicationKey(ctx, id)
-		if err != nil {
-			return fmt.Errorf("received an error retrieving application key %s", err)
-		}
-		last4OfKey = resp.Data.Attributes.GetLast4()
+	resp, _, err := apiInstances.GetKeyManagementApiV2().GetCurrentUserApplicationKey(ctx, id)
+	if err != nil {
+		return fmt.Errorf("received an error retrieving application key %s", err)
 	}
-	if expectedKey[len(expectedKey)-4:] != last4OfKey {
+	actualKey := resp.Data.Attributes.GetKey()
+	if expectedKey != actualKey {
 		return fmt.Errorf("application key value does not match")
 	}
 	return nil
