+++ docs/resources/application_key.md
@@ -17,27 +17,6 @@ Provides a Datadog Application Key resource. This can be used to create and mana
 resource "datadog_application_key" "foo" {
   name = "foo-application"
 }
-
-# Create a new Datadog Service Account Application Key
-# Source a role
-data "datadog_role" "ro_role" {
-  filter = "Datadog Read Only Role"
-}
-
-# Create a new Datadog service account
-resource "datadog_user" "foo" {
-  email = "new@example.com"
-
-  roles = [data.datadog_role.ro_role.id]
-
-  service_account = true
-}
-
-resource "datadog_application_key" "bar" {
-  name = "bar-application"
-
-  service_account = datadog_user.foo.id
-}
 ```
 
 <!-- schema generated by tfplugindocs -->
@@ -46,10 +25,6 @@ resource "datadog_application_key" "bar" {
 ### Required
 
 - `name` (String) Name for Application Key.
-
-### Optional
-
-- `service_account` (String) ID of a service account that owns the Application Key.
 
 ### Read-Only
