+++ docs/resources/user.md
@@ -24,17 +24,6 @@ resource "datadog_user" "foo" {
 
   roles = [data.datadog_role.ro_role.id]
 }
-
-# Create a new Datadog service account
-resource "datadog_user" "bar" {
-  email = "new@example.com"
-
-  name = "Service Account Bar"
-
-  roles = [data.datadog_role.ro_role.id]
-
-  service_account = true
-}
 ```
 
 <!-- schema generated by tfplugindocs -->
@@ -50,7 +39,6 @@ resource "datadog_user" "bar" {
 - `name` (String) Name for user.
 - `roles` (Set of String) A list a role IDs to assign to the user.
 - `send_user_invitation` (Boolean) Whether an invitation email should be sent when the user is created.
-- `service_account` (Boolean) Whether the user is a service account.
 
 ### Read-Only
