+++ examples/resources/datadog_user/resource.tf
@@ -9,14 +9,3 @@ resource "datadog_user" "foo" {
 
   roles = [data.datadog_role.ro_role.id]
 }
-
-# Create a new Datadog service account
-resource "datadog_user" "bar" {
-  email = "new@example.com"
-
-  name = "Service Account Bar"
-
-  roles = [data.datadog_role.ro_role.id]
-
-  service_account = true
-}
